+++ pdfminer/arcfour.py
@@ -22,14 +22,14 @@ class Arcfour:
     def process(self, data: bytes) -> bytes:
         (i, j) = (self.i, self.j)
         s = self.s
-        r = b""
-        for c in iter(data):
+        r = bytearray(len(data))
+        for n, c in enumerate(data):
             i = (i + 1) % 256
             j = (j + s[i]) % 256
             (s[i], s[j]) = (s[j], s[i])
             k = s[(s[i] + s[j]) % 256]
-            r += bytes((c ^ k,))
+            r[n] = c ^ k
         (self.i, self.j) = (i, j)
-        return r
+        return bytes(r)
 
     encrypt = decrypt = process